+++ NEWS.md
@@ -1,5 +1,8 @@
 # rsconnect 0.8.30 (development version)
 
+* `deployApp(appPrimaryDoc)` has been deprecated since carefully analysis 
+  revealed that it did the same job as `recordDir`.
+
 * The `rsconnect.pre.deploy` and `rsconnect.post.deploy` hooks are now always
   called with the content directory, not sometimes the path to a specific file
   (#696).